feat: handled new login flow
