+++ pages/index.tsx
@@ -1,125 +1,116 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 
+import {Session} from 'next-auth';
 import {getSession} from 'next-auth/react';
 import getConfig from 'next/config';
-import dynamic from 'next/dynamic';
 import Head from 'next/head';
-import {useRouter} from 'next/router';
 
-import AlertComponent from 'src/components/atoms/Alert/Alert.component';
-import ShowIf from 'src/components/common/show-if.component';
-import {LoginLayout} from 'src/components/template/Login';
-import {MobileLayout} from 'src/components/template/Login';
-import {useAlertHook} from 'src/hooks/use-alert.hook';
-import {WalletTypeEnum} from 'src/interfaces/wallet';
+import {Timeline} from 'components/Timeline/Timeline.layout';
+import {SearchBoxContainer} from 'components/atoms/Search/SearchBoxContainer';
+import {NavbarComponent} from 'src/components/Mobile/Navbar/Navbar';
+import {RichTextContainer} from 'src/components/Richtext/RichTextContainer';
+import {AppStatusBanner} from 'src/components/common/Banner';
+import {TippingSuccess} from 'src/components/common/Tipping/render/Tipping.success';
+import {DefaultLayout} from 'src/components/template/Default/DefaultLayout';
 import {initialize} from 'src/lib/api/base';
+import {healthcheck} from 'src/lib/api/healthcheck';
+import {getServer} from 'src/lib/api/server';
 import i18n from 'src/locale';
-import {fetchNetwork} from 'src/reducers/user/actions';
+import {fetchAvailableToken} from 'src/reducers/config/actions';
+import {fetchExchangeRates} from 'src/reducers/exchange-rate/actions';
+import {fetchFriend} from 'src/reducers/friend/actions';
+import {countNewNotification} from 'src/reducers/notification/actions';
+import {
+  setAnonymous,
+  fetchConnectedSocials,
+  fetchUser,
+  fetchUserExperience,
+  fetchUserWallets,
+  fetchNetwork,
+} from 'src/reducers/user/actions';
 import {wrapper} from 'src/store';
 import {ThunkDispatchAction} from 'src/types/thunk';
 
-const Login = dynamic(() => import('src/components/Login/Login'), {
-  ssr: false,
-});
-const OnBoardingContainer = dynamic(
-  () => import('src/components/Mobile/OnBoardingView/OnBoardingView.container'),
-  {
-    ssr: false,
-  },
-);
-
 const {publicRuntimeConfig} = getConfig();
 
-const description = i18n.t('Login.Description');
-
-type IndexPageProps = {
-  mobile: boolean;
-  redirectAuth: WalletTypeEnum | null;
+type HomePageProps = {
+  session: Session;
+  logo: string;
 };
 
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export default function Index(props: IndexPageProps) {
-  const {mobile, redirectAuth} = props;
-
-  const {query} = useRouter();
-
-  const {showAlert} = useAlertHook();
-
-  useEffect(() => {
-    if (query.error) {
-      showAlert({
-        message: i18n.t('Login.Alert.Message'),
-        severity: 'error',
-        title: i18n.t('Login.Alert.Title'),
-      });
-    }
-  }, [query.error]);
-
+const Index: React.FC<HomePageProps> = props => {
   return (
-    <>
+    <DefaultLayout isOnProfilePage={false} {...props}>
       <Head>
-        <meta name="description" content={description}></meta>
-        <meta property="og:title" content={publicRuntimeConfig.appName} key="ogtitle" />
-        <meta property="og:description" content={description} key="ogdesc" />
-        <title>{publicRuntimeConfig.appName}</title>
+        <title>{i18n.t('Home.Title', {appname: publicRuntimeConfig.appName})}</title>
       </Head>
 
-      <ShowIf condition={!mobile}>
-        <LoginLayout>
-          <Login redirectAuth={redirectAuth} />
-          <AlertComponent />
-        </LoginLayout>
-      </ShowIf>
+      <NavbarComponent {...props} />
 
-      <ShowIf condition={mobile}>
-        <MobileLayout>
-          <OnBoardingContainer redirectAuth={redirectAuth} />
-        </MobileLayout>
-      </ShowIf>
-    </>
+      <SearchBoxContainer hidden={true} />
+
+      <RichTextContainer />
+
+      <Timeline />
+
+      <TippingSuccess />
+
+      <AppStatusBanner />
+    </DefaultLayout>
   );
-}
+};
 
 export const getServerSideProps = wrapper.getServerSideProps(store => async context => {
-  const {req, query} = context;
-  const {headers} = req;
+  const {req} = context;
+
   const dispatch = store.dispatch as ThunkDispatchAction;
 
-  let mobile = false;
-  let redirectAuth: string | null = null;
+  const available = await healthcheck();
 
-  if (query.auth) {
-    redirectAuth = Array.isArray(query.auth) ? query.auth[0] : query.auth;
-  }
-
-  if (typeof window === 'undefined' && headers['user-agent']) {
-    const UAParser = eval('require("ua-parser-js")');
-    const parser = new UAParser();
-    const device = parser.setUA(headers['user-agent']).getDevice();
-
-    if (device.type === 'mobile') {
-      mobile = true;
-    }
-  }
-
-  const session = await getSession(context);
-
-  if (session) {
+  if (!available) {
     return {
       redirect: {
-        destination: '/home',
+        destination: '/maintenance',
         permanent: false,
       },
     };
   }
 
-  initialize();
-  await dispatch(fetchNetwork());
+  const session = await getSession(context);
+
+  const anonymous = Boolean(session?.user.anonymous);
+  const userId = session?.user.address;
+
+  initialize({cookie: req.headers.cookie}, anonymous);
+
+  if (anonymous || !userId) {
+    await dispatch(setAnonymous('anonymous'));
+  } else {
+    await dispatch(fetchUser(userId));
+
+    await Promise.all([
+      dispatch(fetchConnectedSocials()),
+      dispatch(fetchUserWallets()),
+      dispatch(countNewNotification()),
+      dispatch(fetchFriend()),
+    ]);
+  }
+
+  await Promise.all([
+    dispatch(fetchAvailableToken()),
+    dispatch(fetchNetwork()),
+    dispatch(fetchExchangeRates()),
+  ]);
+
+  await dispatch(fetchUserExperience());
+  const data = await getServer();
 
   return {
     props: {
-      mobile,
-      redirectAuth,
+      session,
+      logo: data.images.logo_banner,
     },
   };
 });
+
+export default Index;
